Migrate to v2 SDK
This is a major update to the terraform SDK.  The details can be found
here:  https://www.terraform.io/docs/extend/guides/v2-upgrade-guide.html
Among some of the benefits are:
- improved handling of context, in particular more reliable timeouts and
  cancellation
- improved error reporting and handling

To gain the above benefits new context-away CRUD function signatures
need to be used.  That migration can be done in subsequent PRs.

The majority of the migration was performed with the tf-sdk-migrator
using a two-step process:
- migrate from terraform to v1 SDK
- migrate from v1 SDK to v2 SDK

There were a few manual changes that had to made:
- we needed to manually switch to the schema.Provider type
- we had to update the init() of test provider resources and the
  TestAccProvider test
- Schema maps with element type of Resource aren't supported yet by the
  v2 SDK, this impacted schema keys of "profile" and "limits" under the
  "project" resource.  We used a recommended workaround of declaring a
  list with just one element for each.

We also uncovered a couple of typo bugs in the network_resource code,
the wrong keys were being used in d.Set - n.Description and n.LocationID
instead of nDescription and nLocationID.

The acceptance tests pass.