HPE GreenLake Terraform Provider import prep - 1
In this PR we make the first set of changes to allow the poc HPE
Greenlake Terraform provider to import the Quake Terraform code from
this repo and expose the resources and data-sources as "hpegl_quake_*"
resources and data-sources.  We're using the same method as that of the
Azurerm provider to do this.  We've:
- exported the quake terraform provider resource and data-source
  functions
- added pkg/terraform/registration/registration.go which can be imported
  by the Greenlake Terraform provider
- registration.go defines a struct which exposes three receivers that
  conform to the ServiceRegistration interface used by the GreenLake
  provider
- SupportedDataSources() returns a map[string]*schema.Resource which
  maps the new hpegl_quake_* data-source names to the corresponding
  quake data-source functions
- SupportedResources() returns a map[string]*schema.Resource which maps
  the new hpegl_quake_* resource names to the corresponding quake
  resource functions