HPE GreenLake Terraform Provider import prep - 3
In this PR we move the Config struct and its related functions and
structs contained in config.go, gl_creds.go, legacy_creds.go and
legacy_creds_test.go to pkg/terraform/configuration.  This is so that
Config can be exported to other repos, in particular the HPE GreenLake
terraform provider repo.  We've exported the Config receivers, and also
exported some more elements of the Config struct.  We've adjusted the
terraform code and tests to take account of this move.