HPE GreenLake Terraform Provider import prep - 5 (#78)
* HPE GreenLake Terraform Provider import prep - 5

In this PR we adapt the quake terraform code to expect that meta might
be of type map[string]interface{} or *configuration.Config.  The map
structure will be passed from the GL provider.  If we get this map
structure we look for *configuration.Config at the key returned by
configuration.KeyInGLClientMap().  If we can't find
*configuration.Config at either of the two possibilities we return an
error.  The logic for this is contained in a new utility function
getConfigFromMeta().  We've added a unit-test for this new function.

* Mark's changes