Update hpegl Registration interface (#83)
In this PR we update the pkg/terraform/registration interface definition
to include a ProviderSchemaEntry() implementation.  We also change the
return-value of the Name() function to "bmaas".  The
ProviderSchemaEntry() function is used by the hpegl provider to include
a "bmaas" block in the hpegl{} stanza that allows a user to specify the
project-id, rest-url and space-name for inclusion in a .gltform file.
The values of these fields can also be set by env-vars.  These additions
should make it easier for a user to interact with the hpegl provider
when managing bmaas/quake resources.