Simple terraform schema fix (#9)
* Simple terraform schema fix

We're making use of terraform-plugin-docs to automatically generate
hpegl provider documentation.  This tooling has exposed an inconsistency
in the schema definition, which has required a one-line change to
resource_project.go

* Fix copyright statement

* Switch from TypeList to TypeSet

It turns out that the error being generated by tfplugindocs can be fixed
by changing TypeList to TypeMap for pLimits and for pProfile.  TypeMap
wasn't supported by the v2.0 SDK when we first migrated to it.

* Revert TypeList to TypeMap change

It turns out that nested TypeMaps - i.e. Maps with Resources within Maps
- still aren't supported by the v2.0 SDK.  Instead we've changed the
pLimits elem fields to "Optional: true" from "Computed: true".  Looking
at the code those limits are expected to be supplied by the user, and
thus aren't computed.