Remove vendor directory (#16)
* Remove vendor directory

In this PR we remove the vendor directory.  We've added a target to each
Makefile to create the vendor directory as necessary.

We've added the vendor directory to .gitignore

We've also restructured secret-scanning in .circleci/config.yml to
conform with standard practice across hpe-hcss:
- in its own workflow
- running against /src and not ~/project

* Adding .gitleaks-local.toml to ignore the vendor directory

* Fix gitleaks

* Adding .copyrightignore

* Fix gitleaks regex

* Fix regex 2

* Temporarily remove secret-scanning

We cannot get past secret-scanning reporting leaks from vendor.  So
removing and then we'll reinstate when vendor is gone.

* Removing .gitleaks-local.toml

* Removing scanner from Dockerfile.circleci